Trying to refactor too much
