+++ libsignal-service-actix/src/lib.rs
@@ -6,5 +6,3 @@ pub mod websocket;
 pub mod prelude {
     pub use crate::push_service::*;
 }
-
-pub mod provisioning;
